NDRS-833: Extract `Signature` to a file.
